\name{get_mesh_agree_pair_long}
\alias{get_mesh_agree_pair_long}
\title{asdf}
\description{asdf.}
\usage{
get_mesh_agree_pair_long(x)
}
\arguments{
  \item{x}{asdf.}
}
\details{asdf.}
\value{asdf.}
\seealso{
\code{\link[MeshAgreement]{read_mesh}},
\code{\link[MeshAgreement]{print_mesh}}
}
\examples{
1
}
