\name{get_mesh_metro}
\alias{get_mesh_metro}
\title{Distances between two 3D meshes}
\description{Calculate distances between two 3D meshes. Used as a basis for distance-based agreement measures such as the Hausdorff distance or the average surface distance.}
\usage{get_mesh_metro(x, n_samples=1000L, chop=TRUE)}
\arguments{
  \item{x}{\code{list} of lists 3D meshes, e.g., as generated by \code{\link[MeshAgreement]{read_mesh}}.}
  \item{n_samples}{Number of samples to save time in generating distance map between meshes in \code{\link[Rvcg]{vcgMetro}}. Set to \code{0} to avoid sampling and use all vertices.}
  \item{chop}{\code{logical}. Delete components from output of \code{\link[Rvcg]{vcgMetro}} that will not be used later on.}
}
\details{Thin front-end for \code{\link[Rvcg]{vcgMetro}}. Use \code{\link[MeshAgreement]{get_mesh_metro_pair}} to process a single mesh pair. Post-processing for agreement measures is done in \code{\link[MeshAgreement]{get_mesh_agree}}.}
\seealso{
\code{\link[Rvcg]{vcgMetro}},
\code{\link[MeshAgreement]{get_mesh_metro_pair}},
\code{\link[MeshAgreement]{get_mesh_agree}}
}
\examples{
metroL <- get_mesh_metro(data_heart_meshL)
# Hausdorff distance
get_HD_max <- function(x) {
    max(c(x$ForwardSampling$maxdist,
          x$BackwardSampling$maxdist))
}

Map(get_HD_max, metroL)
}
