\name{get_mesh_ui}
\alias{get_mesh_ui}
\title{Union and intersection for a list of mesh pairs}
\description{Calculates the union and intersection for a list of pairs of mesh objects.}
\usage{
get_mesh_ui_pair(x)
}
\arguments{
  \item{x}{\code{list} of lists with two mesh objects each as generated by \code{\link[MeshAgreement]{get_mesh_pairs}}.}
}
\details{Front-end for \code{\link[Boov]{MeshesUnion}} and \code{\link[Boov]{MeshesIntersection}} which are based on the CGAL library for computational geometry. Use \code{\link[MeshAgreement]{get_mesh_ui_pair}} to get the union and intersection for a single pair of meshes.}
\value{\code{list} of lists with two \code{\link[rgl]{mesh3d}} objects each and the name of the comparison.}
\references{
\url{https://www.cgal.org/}
}
\seealso{
\code{\link[MeshAgreement]{get_mesh_ui_pair}},
\code{\link[Boov]{MeshesUnion}},
\code{\link[Boov]{MeshesIntersection}}
}
\examples{
pairL <- get_mesh_pairs(data_heart_meshL)
uiL <- get_mesh_ui(pairL)
uiL
}
