\name{get_mesh_ui_pair}
\alias{get_mesh_ui_pair}
\title{Union and intersection for a single pair of 3D meshes}
\description{Calculates the union and intersection of a single pair of 3D mesh objects.}
\usage{
get_mesh_ui_pair(x)
}
\arguments{
  \item{x}{\code{list} of two mesh objects as generated by \code{\link[MeshAgreement]{get_mesh_pairs}}.}
}
\details{Front-end for \code{\link[Boov]{MeshesUnion}} and \code{\link[Boov]{MeshesIntersection}} which are based on the CGAL library for computational geometry. Use \code{\link[MeshAgreement]{get_mesh_ui}} to get the respective union and intersection for all pairs from a list of 3D meshes.}
\value{\code{list} of two \code{\link[rgl]{mesh3d}} objects with the name of the comparison.}
\references{
\url{https://www.cgal.org/}
}
\seealso{
\code{\link[MeshAgreement]{get_mesh_ui}},
\code{\link[Boov]{MeshesUnion}},
\code{\link[Boov]{MeshesIntersection}}
}
\examples{
pairL <- get_mesh_pairs(data_heart_meshL)
ui <- get_mesh_ui_pair(pairL[[1]])
ui
}
