\name{mesh_list_to_observer_list}
\alias{mesh_list_to_observer_list}
\title{Turn list of meshes into list of observer/mesh list}
\description{Turns a list of meshes into a list of observers, each with a list of meshes.}
\usage{
mesh_list_to_observer_list(x)
}
\arguments{
  \item{x}{\code{list} of meshes as generated by \code{\link[MeshAgreement]{read_mesh_obs}}.}
}
\details{Each mesh of the mesh list is assigned to a different observer. As a result, there will be as many observers as meshes, each observer with one mesh.}
\value{\code{list} of observer/mesh objects as generated by \code{\link[MeshAgreement]{read_mesh}}.}
\seealso{
\code{\link[MeshAgreement]{read_mesh}},
\code{\link[MeshAgreement]{read_mesh_obs}}
}
\examples{
meshL <- data_heart_obsL[["Obs01"]]
obsL <- mesh_list_to_observer_list(meshL)
}
