\name{MeshAgreement-package}
\alias{MeshAgreement-package}
\alias{MeshAgreement}
\docType{package}
\title{Agreement measures for 3D meshes}
\description{
Calculates distance-based and volume-overlap-based
agreement measures for 3D meshes. Overall agreement for
a set of meshes is calculated as the aggregate agreement
for all pairwise comparisons.
Includes web-based graphical user interface.
}
\details{
\tabular{ll}{
Package: \tab MeshAgreement\cr
Type: \tab Package\cr
Version: \tab 0.1.0.999\cr
Date: \tab 2022-10-28\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Daniel Wollschlaeger and Heiko Karle

Maintainer: Daniel Wollschlaeger <wollschlaeger@uni-mainz.de>
}
\references{}
\keyword{package}
\examples{
1
}
