\name{data_heart_meshL}
\alias{data_heart_meshL}
\docType{data}
\title{asdf}
\description{asdf}
\usage{data(data_heart_meshL)}
\format{asdf}
\details{Data courtesy of Department of Radiation Oncology, University Medical Center Mainz, Germany.}
\seealso{
\code{\link[MeshAgreement]{read_mesh}},
\code{\link[MeshAgreement]{print_mesh}}
}
\examples{
print_mesh(data_heart_mesh)
}
\keyword{datasets}
