\name{get_mesh_agree}
\alias{get_mesh_agree}
\title{Agreement measures for pairs from a list of 3D meshes}
\description{Calculates distance-based as well as volume-overlap-based agreement measures in each possible pairwise comparison for a list of 3D meshes.}
\usage{
get_mesh_agree(x, n_samples=1000L, chop=TRUE)
}
\arguments{
  \item{x}{\code{list} of 3D meshes as generated by \code{\link[MeshAgreement]{read_mesh}}.}
  \item{n_samples}{Number of samples to save time in generating the distance map between two meshes in \code{\link[Rvcg]{vcgMetro}}. Set to \code{0} to avoid sampling and use all vertices.}
  \item{chop}{\code{logical}. Delete components from the output of \code{\link[Rvcg]{vcgMetro}} that will not be used later on.}
}
\details{Currently, no mesh can be designated gold standard. \code{\link[MeshAgreement]{get_mesh_agree_pair}} calculates the agreement for a single pair of meshes. \code{\link[MeshAgreement]{get_mesh_metro}} calculates the distance measures. \code{\link[MeshAgreement]{get_mesh_ui}} calculates the mesh union and intersection.}
\value{
\code{data.frame} with these variables
\item{mesh1}{Name of the first mesh of the pair.}
\item{mesh2}{Name of the second mesh of the pair.}
\item{DCOM}{Euclidean distance between the respective center of mass of both meshes.}
\item{HD_max}{Hausdorff distance - worst case, maximum of both directed Hausdorff distances.}
\item{HD_avg}{Hausdorff distance - average, mean of both directed Hausdorff distancess.}
\item{HD_95}{95th percentile Hausdorff distance - mean of both 0.95 quantiles of directed distances.}
\item{ASD}{Average symmetric surface distance.}
\item{RMSD}{Root mean squared symmetric surface distance.}
\item{JSC}{Jaccard similarity coefficient.}
\item{DSC}{Dice similarity coefficient.}
}
\references{
Fotina et al. Critical discussion of evaluation parameters for inter-observer variability in target definition for radiation therapy. Strahlenther Onkol 2012; 188: 160-167.

Hanna  et al. Geometrical Analysis of Radiotherapy Target Volume Delineation: a Systematic Review of Reported Comparison Methods. Clin Oncol 2010; 22, 515-525.

Heimann et al. Comparison and Evaluation of Methods for Liver Segmentation From CT Datasets. IEEE Trans Med Imaging 2009; 28: 1251-1265.

Sherer et al. Metrics to evaluate the performance of auto-segmentation for radiation treatment planning: A critical review. Radiother Oncol 2021; 160: 185-191.
}
\seealso{
\code{\link[MeshAgreement]{get_mesh_agree_pair}},
\code{\link[MeshAgreement]{get_mesh_metro}},
\code{\link[MeshAgreement]{get_mesh_ui}}
}
\examples{
get_mesh_agree(data_heart_meshL)
}
