\name{get_mesh_metro}
\alias{get_mesh_metro}
\title{asdf}
\description{asdf.}
\usage{get_mesh_metro(x, n_samples=1000L, chop=TRUE)}
\arguments{
  \item{x}{asdf.}
  \item{n_samples}{asdf.}
  \item{chop}{asdf.}
}
\details{asdf.}
\seealso{
\code{\link[MeshAgreement]{read_mesh}},
\code{\link[MeshAgreement]{print_mesh}}
}
\examples{
1
}
