\name{get_mesh_metro_pair}
\alias{get_mesh_metro_pair}
\title{Distances between two 3D meshes}
\description{Calculate distances between two 3D meshes. Used as a basis for distance-based agreement measures such as the Hausdorff distance or the average surface distance.}
\usage{get_mesh_metro_pair(x, n_samples=1000L, chop=TRUE)}
\arguments{
  \item{x}{\code{list} of two 3D meshes that should be compared, e.g., a single component of the list generated by \code{\link[MeshAgreement]{get_mesh_pairs}}.}
  \item{n_samples}{Number of samples to save time in generating distance map between meshes in \code{\link[Rvcg]{vcgMetro}}. Set to \code{0} to avoid sampling and use all vertices.}
  \item{chop}{\code{logical}. Delete components from output of \code{\link[Rvcg]{vcgMetro}} that will not be used later on.}
}
\details{Thin front-end for \code{\link[Rvcg]{vcgMetro}}. Use \code{\link[MeshAgreement]{get_mesh_metro}} to process a list of mesh pairs. Post-processing for agreement measures is done in \code{\link[MeshAgreement]{get_mesh_agree_pair}}.}
\seealso{
\code{\link[Rvcg]{vcgMetro}},
\code{\link[MeshAgreement]{get_mesh_metro}},
\code{\link[MeshAgreement]{get_mesh_agree_pair}}
}
\examples{
pairL <- get_mesh_pairs(data_heart_meshL)
metro <- get_mesh_metro_pair(pairL[[1]])
# Hausdorff distance
get_HD_max <- function(x) {
    max(c(x$ForwardSampling$maxdist,
          x$BackwardSampling$maxdist))
}

get_HD_max(metro)
}
