\name{get_mesh_pairs}
\alias{get_mesh_pairs}
\title{Get all pairs from list of 3D meshes}
\description{Generates all possible pairs based on a list of 3D meshes to enable post-processing with pairwise agreement measures.}
\usage{
get_mesh_pairs(x, sep=" <-> ")
}
\arguments{
  \item{x}{\code{list} of mesh objects as generated by \code{\link[MeshAgreement]{read_mesh}}.}
  \item{sep}{\code{character}. The string that indicates the separator between the first and second element in naming each generated pair of meshes.}
}
\details{Preparation for postprocessing with pairwise agreement measures using \code{\link[MeshAgreement]{get_mesh_agree_pair}}.}
\value{\code{list} of lists with two \code{mesh3d} objects each.}
\seealso{
\code{\link[MeshAgreement]{read_mesh}},
\code{\link[MeshAgreement]{print_mesh}}
}
\examples{
pairL <- get_mesh_pairs(data_heart_meshL)
}
