\name{get_mesh_ui}
\alias{get_mesh_ui}
\title{Union and intersection for pairs from a list of 3D meshes}
\description{Calculates the respective union and intersection for all possible pairs from a list of 3D mesh objects.}
\usage{
get_mesh_ui(x)
}
\arguments{
  \item{x}{\code{list} of 3D meshes as generated by \code{\link[MeshAgreement]{read_mesh}}.}
}
\details{Front-end for \code{\link[Boov]{MeshesUnion}} and \code{\link[Boov]{MeshesIntersection}} which are based on the CGAL library for computational geometry. Use \code{\link[MeshAgreement]{get_mesh_ui_pair}} to get the union and intersection for a single pair of meshes.}
\value{\code{list} of lists with two mesh objects each, and the name of the comparison.}
\references{
\url{https://www.cgal.org/}
}
\seealso{
\code{\link[MeshAgreement]{get_mesh_ui_pair}},
\code{\link[Boov]{MeshesUnion}},
\code{\link[Boov]{MeshesIntersection}}
}
\examples{
uiL <- get_mesh_ui(data_heart_meshL)
uiL
}
