\name{read_mesh_one}
\alias{read_mesh_one}
\title{asdf}
\description{asdf.}
\usage{read_mesh_one(x), name, reconstruct=TRUE}
\arguments{
  \item{x}{asdf.}
  \item{name}{asdf.}
  \item{reconstruct}{asdf.}
}
\details{asdf.}
\seealso{
\code{\link[MeshAgreement]{read_mesh}},
\code{\link[MeshAgreement]{print_mesh}}
}
\examples{
1
}
